requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.13.4
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
DAWG2-Python==0.9.0
Django==5.2.3
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
exceptiongroup==1.3.0
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
preshed==3.0.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
pymorphy3==2.0.4
pymorphy3-dicts-ru==2.4.417150.4580142
PyMuPDF==1.26.1
pytest==8.4.1
pytest-django==4.11.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
rich==14.0.0
ru_core_news_md @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_md-3.8.0/ru_core_news_md-3.8.0-py3-none-any.whl#sha256=43f456d40e4b70726874e01ce01fce32f50089ba98a2660292e9a6c63a06ebe9
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sympy==1.14.0
textdistance==4.6.3
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
tqdm==4.67.1
transformers==4.53.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
